Add files via upload
Lets calculate for each cell of the initial array the number of queries that cover this cell. It’s claimed, that we should associate the cell with bigger value to the cell with bigger value in initial array.
For this purpose one can use different data structures, which support segment modifications (segment tree, Cartesian tree and so on). But there exists much easier method.
Let’s create some array d. When we have query [li, ri,] we should increase value d[li] by 1 and decrease value d[ri + 1] by 1. In such a tricky way we increase all elements in segment [li;ri] by 1 After processing all of the queries we need to make a loop, which visit every element of array d. In this loop we can easily calculate all elements of array b.